% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nFactors.R
\docType{package}
\name{nFactors}
\alias{nFactors-package}
\alias{nFactors}
\title{nFactors: Number of factor or components to retain in a factor analysis}
\description{
A package for determining the number of factor or components to retain in a factor analysis.
The methods are all based on eigenvalues.
}
\references{
Raiche, G., Walls, T. A., Magis, D., Riopel, M. and Blais, J.-G. (2013). Non-graphical solutions
for Cattell's scree test. Methodology, 9(1), 23-29.
}
\author{
Gilles Raiche \cr Centre sur les Applications des Modeles de
Reponses aux Items (CAMRI) \cr Universite du Quebec a Montreal\cr
\email{raiche.gilles@uqam.ca}
}
